apps/api/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
httpx>=0.27,<1
pydantic>=2,<3
pydantic-settings>=2,<3
python-jose[cryptography]>=3.3.0,<4
cryptography>=42,<44
beautifulsoup4>=4.12,<5
html2text>=2020.1.16,<2025
pywebpush>=1.14.0,<2
jsonschema>=4.22.0,<5

